utils: pass LANG=C to subprocess.Popen
ksc stops as Invalid architecture even if the architecture of the module is x86_64.

  $ /usr/share/ksc/ksc.py -k ./foo.ko
  ./foo.ko: Invalid architecture. (only aarch64, i686, ppc64, ppc64le, s390x, x86_64 are supported)
  $

  $ file foo.ko
  foo.ko: ELF 64-bit LSB relocatable, x86-64, version 1 (SYSV),
  BuildID[sha1]=7b22c4fe523acd9e290011fdb3f5d9ff9819ce1f, with debug_info,
  not stripped

That is because ksc cannot edit the result of readelf command
if LANG environment isn't English family, for example LANG=ja_JP.UTF-8.

Pass LANG=C to subprocess.Popen so that ksc can edit the result
of readelf command.

Signed-off-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>