config: change config semantics
1. Default placeholder values are now deprecated

Previously, ksc would allow placeholder names such as user@redhat.com,
partner-name, and partner-group to stand for user, partner, and
partnergroup fields, respectively, and if provided, ksc would query
for actual values at runtime. This is not suitable for batch exection
and is now deprecated.

If found, submission is cancelled and user is asked to fix ksc config.

2. New mailing list section is added.

3. Users are allowed to use ksc w/o partner/partnergroup

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>